PlasmaDispersionFunctions.jl Documentation

PlasmaDispersionFunctions.PlasmaDispersionFunctions — Module

Calculate the plasma dispersion function

principal(n, x) = x.^n .* exp(-x.^2) / √π
integrande(z, n, x) = P(x, n) / (x - z)
residue(z, n, x) = (imag(z) < 0 ? 2 : imag(z) == 0 ? 1 : 0) * im * π * principal(n, z)
Z(z, n) = ∫integrand(z, n, x)dx + residue(z, n, x)

Example

  plasma_dispersion_function(0.0) ≈ im * √π

PlasmaDispersionFunctions.plasma_dispersion_function — Method

@docs plasmadispersionfunction(z::T,power::Unsigned=UInt64(0),Z₋₁=missing)

Return the value of the plasma dispersion function This implementation includes the residue, which is easy to verify because Z(0) = im sqrt(π).

...

Arguments

z::T: the argument to the plasma dispersion function

power::Unsigned=UInt64(0): the moment of the integral

Z₋₁=missing: (optional) the value of the plasma dispersion function for

power-1. If already calculated, passing this in saves some time. ...

Example

  plasma_dispersion_function(1.0)

References

[1] D. Fried and S. D. Conte, "The Plasma Dispersion Relation", Elsevier, 1961 [2] S.D. Baalrud, Phys. Plasmas 20, 012118 (2013) and put ν = -Inf [3] M. Sampoorna et al., Generalized Voigt functions and their derivatives, Journal of Quantitative Spectroscopy & Radiative Transfer (2006), doi:10.1016/j.jqsrt.2006.08.011

Index

PlasmaDispersionFunctions.PlasmaDispersionFunctions

PlasmaDispersionFunctions.plasma_dispersion_function